refactor: Extract ElasticConector from interpreter
